+++ src/main/java/com/connectnow/dao/MemberDao.java
@@ -4,6 +4,6 @@ import com.connectnow.entity.MemberEntity;
 
 import java.math.BigInteger;
 
-public interface MemberDao extends GenericDao<BigInteger, MemberEntity> {
-    MemberEntity findOneByChatBoxIdAndUserId(BigInteger chatBoxId, BigInteger userId);
+public interface MemberDao extends GenericDao<Long, MemberEntity> {
+    MemberEntity findOneByChatBoxIdAndUserId(Long chatBoxId, Long userId);
 }
