+++ src/main/java/com/connectnow/dao/UserDao.java
@@ -2,8 +2,6 @@ package com.connectnow.dao;
 
 import com.connectnow.entity.UserEntity;
 
-import java.math.BigInteger;
-
-public interface UserDao extends GenericDao<BigInteger, UserEntity> {
-    UserEntity findPartnerUserByChatBoxId(BigInteger userId, BigInteger chatBoxId);
+public interface UserDao extends GenericDao<Long, UserEntity> {
+    UserEntity findPartnerUserByChatBoxId(Long userId, Long chatBoxId);
 }
