+++ src/main/java/com/connectnow/dao/impl/MessageDaoImpl.java
@@ -6,25 +6,20 @@ import com.connectnow.paging.Pageable;
 import org.hibernate.Criteria;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
-import org.hibernate.criterion.Criterion;
 import org.hibernate.criterion.Order;
 import org.hibernate.criterion.Restrictions;
-import org.hibernate.transform.Transformers;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 
-import javax.xml.transform.Transformer;
-import java.math.BigInteger;
-import java.util.Collections;
 import java.util.List;
 
 @Repository
-public class MessageDaoImpl extends AbstractDaoImpl<BigInteger, MessageEntity> implements MessageDao {
+public class MessageDaoImpl extends AbstractDaoImpl<Long, MessageEntity> implements MessageDao {
     private Logger logger = LoggerFactory.getLogger(MessageDaoImpl.class);
 
     @Override
-    public List<MessageEntity> findAllByChatBoxId(Pageable pageable, BigInteger chatBoxId) {
+    public List<MessageEntity> findAllByChatBoxId(Pageable pageable, Long chatBoxId) {
         List<MessageEntity> entityList;
 
         Session session = this.getSession();
