+++ src/main/java/com/connectnow/service/ChatBoxService.java
@@ -7,10 +7,10 @@ import com.connectnow.paging.Pageable;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface ChatBoxService extends GenericService<BigInteger, ChatBoxDto> {
-    List<ChatBoxDto> finAllByUserId(Pageable pageable, BigInteger userId);
+public interface ChatBoxService extends GenericService<Long, ChatBoxDto> {
+    List<ChatBoxDto> finAllByUserId(Pageable pageable, Long userId);
 
-    ChatBoxDto findOneByMemberId(BigInteger memberId, BigInteger userId);
+    ChatBoxDto findOneByMemberId(Long memberId, Long userId);
 
     ChatBoxDto updateLastDataByMessage(MessageDto messageDto) throws Exception;
 }
