+++ src/main/java/com/connectnow/service/MemberService.java
@@ -4,8 +4,8 @@ import com.connectnow.dto.MemberDto;
 
 import java.math.BigInteger;
 
-public interface MemberService extends GenericService<BigInteger, MemberDto> {
-    MemberDto updateReadStatusByChatBoxIdAndUserId(BigInteger chatBoxId, BigInteger userId, boolean readStatus) throws Exception;
+public interface MemberService extends GenericService<Long, MemberDto> {
+    MemberDto updateReadStatusByChatBoxIdAndUserId(Long chatBoxId, Long userId, boolean readStatus) throws Exception;
 
-    MemberDto updateReadStatusByMemberId(BigInteger memberId, boolean readStatus) throws Exception;
+    MemberDto updateReadStatusByMemberId(Long memberId, boolean readStatus) throws Exception;
 }
