+++ src/main/java/com/connectnow/service/MessageService.java
@@ -6,6 +6,6 @@ import com.connectnow.paging.Pageable;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface MessageService extends GenericService<BigInteger, MessageDto> {
-    List<MessageDto> findAllByChatBoxId(Pageable pageable, BigInteger chatBoxId);
+public interface MessageService extends GenericService<Long, MessageDto> {
+    List<MessageDto> findAllByChatBoxId(Pageable pageable, Long chatBoxId);
 }
