+++ src/main/java/com/connectnow/service/UserService.java
@@ -6,8 +6,8 @@ import com.connectnow.dto.UserDto;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface UserService extends GenericService<BigInteger, UserDto> {
+public interface UserService extends GenericService<Long, UserDto> {
     UserDto findOneByProviderAndProviderId(String provider, String providerId);
 
-    List<SearchUserDto> findSearchUserByName(String name, BigInteger userId);
+    List<SearchUserDto> findSearchUserByName(String name, Long userId);
 }
