+++ src/test/java/com/connectnow/dao/impl/MemberDaoImplTest.java
@@ -1,7 +1,6 @@
 package com.connectnow.dao.impl;
 
 import com.connectnow.config.SpringWebConfig;
-import com.connectnow.dao.ChatBoxDao;
 import com.connectnow.dao.MemberDao;
 import com.connectnow.entity.MemberEntity;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,9 +9,8 @@ import org.springframework.test.context.testng.AbstractTestNGSpringContextTests;
 import org.springframework.test.context.web.WebAppConfiguration;
 import org.testng.annotations.Test;
 
-import java.math.BigInteger;
-
-import static org.testng.Assert.*;
+import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertNotNull;
 
 @WebAppConfiguration
 @ContextConfiguration(classes = {SpringWebConfig.class})
@@ -22,8 +20,8 @@ public class MemberDaoImplTest extends AbstractTestNGSpringContextTests {
 
     @Test
     public void findOneByChatBoxIdAndUserIdTest() {
-        MemberEntity memberEntity = memberDao.findOneByChatBoxIdAndUserId(BigInteger.valueOf(4), BigInteger.valueOf(6));
+        MemberEntity memberEntity = memberDao.findOneByChatBoxIdAndUserId(4L, 6L);
         assertNotNull(memberEntity);
-        assertEquals(memberEntity.getId(), BigInteger.valueOf(5));
+        assertEquals(memberEntity.getId(), Long.valueOf(5));
     }
 }
